+++ tests/test_polynomial_newton.py
@@ -15,7 +15,7 @@ from numpy.testing import assert_, assert_almost_equal
 
 from minterpy.global_settings import INT_DTYPE
 from minterpy.utils import eval_newton_polynomials
-from minterpy import Grid
+from minterpy import Grid, MultiIndexSet
 
 from minterpy import (
     NewtonPolynomial,
@@ -224,3 +224,167 @@ def test_integrate_over(
     # Assertions
     assert np.isclose(value_nwt, value_lag)
     assert np.isclose(value_nwt, value_can)
+
+
+class TestPolyPolyMultiplication:
+    """All tests related to the poly.-poly. multiplication in the Newton basis.
+    """
+    def test_identity(self, SpatialDimension, PolyDegree, LpDegree):
+        """Test multiplicative identity of polynomial multiplication.
+
+        A polynomial multiplied with a constant polynomial with coefficient
+        value of 1.0 should result in the same polynomial in value.
+        """
+        # Create a Newton polynomial
+        nwt_poly = build_random_newton_polynom(
+            SpatialDimension, PolyDegree, LpDegree
+        )
+
+        # Create a constant identity Newton polynomial
+        exponents = np.zeros((1, SpatialDimension), dtype=np.int_)
+        mi = MultiIndexSet(exponents, LpDegree)
+        coeffs = np.array([1.0])
+        nwt_poly_constant = NewtonPolynomial(mi, coeffs)
+
+        # Multiplication
+        nwt_poly_prod_1 = nwt_poly * nwt_poly_constant
+        nwt_poly_prod_2 = nwt_poly_constant * nwt_poly
+
+        # Assertions
+        assert nwt_poly == nwt_poly_prod_1
+        assert nwt_poly == nwt_poly_prod_2
+
+    def test_identity_multiple_polys(
+        self,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        num_polynomials,
+    ):
+        """Test multiplicative identity with multiple sets of coefficients.
+
+        A polynomial with multiple sets of coefficients multiplied with
+        a constant polynomial with coefficient values of 1.0 should return
+        the same polynomial in value.
+        """
+        # Create a Newton polynomial
+        nwt_poly = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+            num_polynomials,
+        )
+
+        # Create a constant identity Newton polynomial
+        exponents = np.zeros((1, SpatialDimension), dtype=np.int_)
+        mi = MultiIndexSet(exponents, LpDegree)
+        coeffs = np.ones((1, num_polynomials))
+        nwt_poly_constant = NewtonPolynomial(mi, coeffs)
+
+        # Multiplication
+        nwt_poly_prod_1 = nwt_poly * nwt_poly_constant
+        nwt_poly_prod_2 = nwt_poly_constant * nwt_poly
+
+        # Assertions
+        assert nwt_poly == nwt_poly_prod_1
+        assert nwt_poly == nwt_poly_prod_2
+
+    def test_constant_sanity(self, SpatialDimension, PolyDegree, LpDegree):
+        """Test the multiplication with an arbitrary constant polynomial.
+
+        A polynomial multiplied with a constant polynomial should return
+        a polynomial multiplied with a scalar.
+        """
+        # Create a Newton polynomial
+        nwt_poly = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+        )
+
+        # Create a constant identity Newton polynomial
+        exponents = np.zeros((1, SpatialDimension), dtype=np.int_)
+        mi = MultiIndexSet(exponents, LpDegree)
+        coeffs = np.random.rand(1)
+        nwt_poly_constant = NewtonPolynomial(mi, coeffs)
+
+        # Multiplication
+        nwt_poly_prod_1 = nwt_poly * nwt_poly_constant
+        nwt_poly_prod_2 = nwt_poly_constant * nwt_poly
+        nwt_poly_prod_3 = nwt_poly * coeffs[0]
+
+        # Assertions
+        assert nwt_poly_prod_3 == nwt_poly_prod_1
+        assert nwt_poly_prod_3 == nwt_poly_prod_2
+
+    def test_eval(self, SpatialDimension, PolyDegree, LpDegree):
+        """Test the evaluation results of polynomial multiplication."""
+        # Create two Newton polynomials
+        nwt_poly_1 = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+        )
+        # Increase both the spatial dimension and poly. degree so
+        # the polynomial differ
+        nwt_poly_2 = build_random_newton_polynom(
+            SpatialDimension+1,
+            PolyDegree+1,
+            LpDegree,
+        )
+
+        # Generate random test points
+        xx_test = -1 + 2 * np.random.rand(10000, SpatialDimension+1)
+
+        # Compute reference results
+        yy_1 = nwt_poly_1(xx_test[:, :SpatialDimension])
+        yy_2 = nwt_poly_2(xx_test[:, :SpatialDimension+1])
+        yy_ref = yy_1 * yy_2
+
+        # Multiply polynomial
+        nwt_poly_prod = nwt_poly_1 * nwt_poly_2
+        yy_prod = nwt_poly_prod(xx_test)
+
+        # Assertion
+        assert np.allclose(yy_ref, yy_prod)
+
+    def test_eval(
+        self,
+        SpatialDimension,
+        PolyDegree,
+        LpDegree,
+        num_polynomials,
+    ):
+        """Test the evaluation results of poly. multiplication with multiple
+        sets of coefficients.
+        """
+        # Create two Newton polynomials
+        nwt_poly_1 = build_random_newton_polynom(
+            SpatialDimension,
+            PolyDegree,
+            LpDegree,
+            num_polynomials,
+        )
+        # Increase both the spatial dimension and poly. degree so
+        # the polynomial differ
+        nwt_poly_2 = build_random_newton_polynom(
+            SpatialDimension + 1,
+            PolyDegree + 1,
+            LpDegree,
+            num_polynomials,
+        )
+
+        # Generate random test points
+        xx_test = -1 + 2 * np.random.rand(1000, SpatialDimension+1)
+
+        # Compute reference results
+        yy_1 = nwt_poly_1(xx_test[:, :SpatialDimension])
+        yy_2 = nwt_poly_2(xx_test[:, :SpatialDimension+1])
+        yy_ref = yy_1 * yy_2
+
+        # Multiply polynomial
+        nwt_poly_prod = nwt_poly_1 * nwt_poly_2
+        yy_prod = nwt_poly_prod(xx_test)
+
+        # Assertion
+        assert np.allclose(yy_ref, yy_prod)